+++ utils/sql/git/optional/2026_05_13_monomyth_character_buffs_is_clicky_buff.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `character_buffs`
+	ADD COLUMN `is_clicky_buff` TINYINT(3) UNSIGNED NOT NULL DEFAULT 0 AFTER `group_raid_timer_pausable`;